refactor: score 타입 변경 및 유효성 검사 추가 #2
